- change long to intptr_t for 64bit windows compatability - change vmMain arguments back to int. 64bit types are apparently not   needed there. Only the syscall function needs them.

+++ code/cgame/cg_syscalls.c
@@ -28,10 +28,10 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #include "cg_local.h"
 
-static long (QDECL *syscall)( long arg, ... ) = (long (QDECL *)( long, ...))-1;
+static intptr_t (QDECL *syscall)( intptr_t arg, ... ) = (intptr_t (QDECL *)( intptr_t, ...))-1;
 
 
-void dllEntry( long (QDECL  *syscallptr)( long arg,... ) ) {
+void dllEntry( intptr_t (QDECL  *syscallptr)( intptr_t arg,... ) ) {
 	syscall = syscallptr;
 }
